// src/Interfaces/Contact/IContact.tsx
interface IContact {
    statiqueField: {
        id: number
        user: string
        language: string
        fullName: string
        email: string
        phone: string
        linkedin: string
        account: string
        position: string
        address: string
        country: string
        city: string
        postalcode: string
    },
    customFields: []

}
export type { IContact }